+++ HISTORY.rst
@@ -7,7 +7,7 @@ Pending Release
 
 .. Modify the below with new release notes
 
-* Nothing right now
+* Drop Python 2 support, only Python 3.4+ is supported now.
 
 2.4.0 (2018-09-30)
 ------------------